Split component code into multiple files
